+++ lib/ecto/adapter/migration.ex
@@ -16,12 +16,12 @@ defmodule Ecto.Adapter.Migration do
           | {:create, Table.t(), [table_subcommand]}
           | {:create_if_not_exists, Table.t(), [table_subcommand]}
           | {:alter, Table.t(), [table_subcommand]}
-          | {:drop, Table.t()}
-          | {:drop_if_exists, Table.t()}
+          | {:drop, Table.t(), :restrict | :cascade}
+          | {:drop_if_exists, Table.t(), :restrict | :cascade}
           | {:create, Index.t()}
           | {:create_if_not_exists, Index.t()}
-          | {:drop, Index.t()}
-          | {:drop_if_exists, Index.t()}
+          | {:drop, Index.t(), :restrict | :cascade}
+          | {:drop_if_exists, Index.t(), :restrict | :cascade}
 
   @typedoc "All commands allowed within the block passed to `table/2`"
   @type table_subcommand ::