+++ lib/ecto/adapters/tds/connection.ex
@@ -974,9 +974,11 @@ if Code.ensure_loaded?(Tds) do
     # DDL
     alias Ecto.Migration.{Table, Index, Reference, Constraint}
 
+    @creates [:create, :create_if_not_exists]
+    @drops [:drop, :drop_if_exists]
+
     @impl true
-    def execute_ddl({command, %Table{} = table, columns})
-        when command in [:create, :create_if_not_exists] do
+    def execute_ddl({command, %Table{} = table, columns}) when command in @creates do
       prefix = table.prefix
 
       pk_name =
@@ -1009,7 +1011,10 @@ if Code.ensure_loaded?(Tds) do
       ]
     end
 
-    def execute_ddl({command, %Table{} = table}) when command in [:drop, :drop_if_exists] do
+    def execute_ddl({command, %Table{}, :cascade}) when command in @drops,
+      do: error!(nil, "MSSQL does not support `CASCADE` in DROP TABLE commands")
+
+    def execute_ddl({command, %Table{} = table, :restrict}) when command in @drops do
       prefix = table.prefix
 
       [
@@ -1044,8 +1049,7 @@ if Code.ensure_loaded?(Tds) do
       ]
     end
 
-    def execute_ddl({command, %Index{} = index})
-        when command in [:create, :create_if_not_exists] do
+    def execute_ddl({command, %Index{} = index}) when command in @creates do
       prefix = index.prefix
 
       if index.using do
@@ -1124,7 +1128,10 @@ if Code.ensure_loaded?(Tds) do
       ]
     end
 
-    def execute_ddl({command, %Index{} = index}) when command in [:drop, :drop_if_exists] do
+    def execute_ddl({command, %Index{}, :cascade}) when command in @drops,
+      do: error!(nil, "MSSQL does not support `CASCADE` in DROP INDEX commands")
+
+    def execute_ddl({command, %Index{} = index, :restrict}) when command in @drops do
       prefix = index.prefix
 
       [
@@ -1144,8 +1151,10 @@ if Code.ensure_loaded?(Tds) do
       ]
     end
 
-    def execute_ddl({command, %Constraint{} = constraint})
-        when command in [:drop, :drop_if_exists] do
+    def execute_ddl({command, %Constraint{}, :cascade}) when command in @drops,
+      do: error!(nil, "MSSQL does not support `CASCADE` in DROP CONSTRAINT commands")
+
+    def execute_ddl({command, %Constraint{} = constraint, _}) when command in @drops do
       table_name = quote_table(constraint.prefix, constraint.table)
 
       [