+++ lib/ecto/migration/runner.ex
@@ -217,18 +217,18 @@ defmodule Ecto.Migration.Runner do
   end
 
   defp reverse({:create, %Index{} = index}),
-    do: {:drop, index}
+    do: {:drop, index, :restrict}
   defp reverse({:create_if_not_exists, %Index{} = index}),
-    do: {:drop_if_exists, index}
-  defp reverse({:drop, %Index{} = index}),
+    do: {:drop_if_exists, index, :restrict}
+  defp reverse({:drop, %Index{} = index, _}),
     do: {:create, index}
-  defp reverse({:drop_if_exists, %Index{} = index}),
+  defp reverse({:drop_if_exists, %Index{} = index, _}),
     do: {:create_if_not_exists, index}
 
   defp reverse({:create, %Table{} = table, _columns}),
-    do: {:drop, table}
+    do: {:drop, table, :restrict}
   defp reverse({:create_if_not_exists, %Table{} = table, _columns}),
-    do: {:drop_if_exists, table}
+    do: {:drop_if_exists, table, :restrict}
   defp reverse({:rename, %Table{} = table_current, %Table{} = table_new}),
     do: {:rename, table_new, table_current}
   defp reverse({:rename, %Table{} = table, current_column, new_column}),
@@ -242,9 +242,9 @@ defmodule Ecto.Migration.Runner do
   # It is not a good idea to reverse constraints because
   # we can't guarantee data integrity when applying them back.
   defp reverse({:create_if_not_exists, %Constraint{} = constraint}),
-    do: {:drop_if_exists, constraint}
+    do: {:drop_if_exists, constraint, :restrict}
   defp reverse({:create, %Constraint{} = constraint}),
-    do: {:drop, constraint}
+    do: {:drop, constraint, :restrict}
 
   defp reverse(_command), do: false
 
@@ -361,19 +361,19 @@ defmodule Ecto.Migration.Runner do
     do: "create table if not exists #{quote_name(table.prefix, table.name)}"
   defp command({:alter, %Table{} = table, _}),
     do: "alter table #{quote_name(table.prefix, table.name)}"
-  defp command({:drop, %Table{} = table}),
-    do: "drop table #{quote_name(table.prefix, table.name)}"
-  defp command({:drop_if_exists, %Table{} = table}),
-    do: "drop table if exists #{quote_name(table.prefix, table.name)}"
+  defp command({:drop, %Table{} = table, mode}),
+    do: "drop table #{quote_name(table.prefix, table.name)}#{drop_mode(mode)}"
+  defp command({:drop_if_exists, %Table{} = table, mode}),
+    do: "drop table if exists #{quote_name(table.prefix, table.name)}#{drop_mode(mode)}"
 
   defp command({:create, %Index{} = index}),
     do: "create index #{quote_name(index.prefix, index.name)}"
   defp command({:create_if_not_exists, %Index{} = index}),
     do: "create index if not exists #{quote_name(index.prefix, index.name)}"
-  defp command({:drop, %Index{} = index}),
-    do: "drop index #{quote_name(index.prefix, index.name)}"
-  defp command({:drop_if_exists, %Index{} = index}),
-    do: "drop index if exists #{quote_name(index.prefix, index.name)}"
+  defp command({:drop, %Index{} = index, mode}),
+    do: "drop index #{quote_name(index.prefix, index.name)}#{drop_mode(mode)}"
+  defp command({:drop_if_exists, %Index{} = index, mode}),
+    do: "drop index if exists #{quote_name(index.prefix, index.name)}#{drop_mode(mode)}"
   defp command({:rename, %Table{} = current_table, %Table{} = new_table}),
     do: "rename table #{quote_name(current_table.prefix, current_table.name)} to #{quote_name(new_table.prefix, new_table.name)}"
   defp command({:rename, %Table{} = table, current_column, new_column}),
@@ -387,10 +387,13 @@ defmodule Ecto.Migration.Runner do
     do: "create check constraint #{constraint.name} on table #{quote_name(constraint.prefix, constraint.table)}"
   defp command({:create, %Constraint{exclude: exclude} = constraint}) when is_binary(exclude),
     do: "create exclude constraint #{constraint.name} on table #{quote_name(constraint.prefix, constraint.table)}"
-  defp command({:drop, %Constraint{} = constraint}),
+  defp command({:drop, %Constraint{} = constraint, _}),
     do: "drop constraint #{constraint.name} from table #{quote_name(constraint.prefix, constraint.table)}"
-  defp command({:drop_if_exists, %Constraint{} = constraint}),
+  defp command({:drop_if_exists, %Constraint{} = constraint, _}),
     do: "drop constraint if exists #{constraint.name} from table #{quote_name(constraint.prefix, constraint.table)}"
+
+  defp drop_mode(:restrict), do: ""
+  defp drop_mode(:cascade), do: " cascade"
 
   defp quote_name(nil, name), do: quote_name(name)
   defp quote_name(prefix, name), do: quote_name(prefix) <> "." <> quote_name(name)