+++ lib/mix/ecto_sql.ex
@@ -4,7 +4,7 @@ defmodule Mix.EctoSQL do
   @doc """
   Ensures the given repository's migrations paths exists on the file system.
   """
-  @spec ensure_migrations_paths(Ecto.Repo.t, Keyword.t) :: String.t
+  @spec ensure_migrations_paths(Ecto.Repo.t, Keyword.t) :: [String.t]
   def ensure_migrations_paths(repo, opts) do
     paths = Keyword.get_values(opts, :migrations_path)
     paths = if paths == [], do: [Path.join(source_repo_priv(repo), "migrations")], else: paths