+++ test/ecto/adapters/tds_test.exs
@@ -1301,12 +1301,18 @@ defmodule Ecto.Adapters.TdsTest do
   end
 
   test "drop table" do
-    drop = {:drop, table(:posts)}
+    drop = {:drop, table(:posts), :restrict}
     assert execute_ddl(drop) == ["DROP TABLE [posts]; "]
+
+    drop_cascade = {:drop, table(:posts), :cascade}
+
+    assert_raise ArgumentError, ~r"MSSQL does not support `CASCADE` in DROP TABLE commands", fn ->
+      execute_ddl(drop_cascade)
+    end
   end
 
   test "drop table with prefixes" do
-    drop = {:drop, table(:posts, prefix: :foo)}
+    drop = {:drop, table(:posts, prefix: :foo), :restrict}
     assert execute_ddl(drop) == ["DROP TABLE [foo].[posts]; "]
   end
 
@@ -1412,19 +1418,25 @@ defmodule Ecto.Adapters.TdsTest do
   end
 
   test "drop constraint" do
-    drop = {:drop, constraint(:products, "price_must_be_positive")}
+    drop = {:drop, constraint(:products, "price_must_be_positive"), :restrict}
 
     assert execute_ddl(drop) ==
              [~s|ALTER TABLE [products] DROP CONSTRAINT [price_must_be_positive]; |]
 
-    drop = {:drop, constraint(:products, "price_must_be_positive", prefix: "foo")}
+    drop = {:drop, constraint(:products, "price_must_be_positive", prefix: "foo"), :restrict}
 
     assert execute_ddl(drop) ==
              [~s|ALTER TABLE [foo].[products] DROP CONSTRAINT [price_must_be_positive]; |]
+
+    drop_cascade = {:drop, constraint(:products, "price_must_be_positive"), :cascade}
+
+    assert_raise ArgumentError, ~r"MSSQL does not support `CASCADE` in DROP CONSTRAINT commands", fn ->
+      execute_ddl(drop_cascade)
+    end
   end
 
   test "drop_if_exists constraint" do
-    drop = {:drop_if_exists, constraint(:products, "price_must_be_positive")}
+    drop = {:drop_if_exists, constraint(:products, "price_must_be_positive"), :restrict}
 
     assert execute_ddl(drop) ==
              [
@@ -1434,7 +1446,7 @@ defmodule Ecto.Adapters.TdsTest do
                  "ALTER TABLE [products] DROP CONSTRAINT [price_must_be_positive]; "
              ]
 
-    drop = {:drop_if_exists, constraint(:products, "price_must_be_positive", prefix: "foo")}
+    drop = {:drop_if_exists, constraint(:products, "price_must_be_positive", prefix: "foo"), :restrict}
 
     assert execute_ddl(drop) ==
              [
@@ -1444,6 +1456,12 @@ defmodule Ecto.Adapters.TdsTest do
                  "AND [CONSTRAINT_SCHEMA] = N'foo') " <>
                  "ALTER TABLE [foo].[products] DROP CONSTRAINT [price_must_be_positive]; "
              ]
+
+    drop_cascade = {:drop_if_exists, constraint(:products, "price_must_be_positive"), :cascade}
+
+    assert_raise ArgumentError, ~r"MSSQL does not support `CASCADE` in DROP CONSTRAINT commands", fn ->
+      execute_ddl(drop_cascade)
+    end
   end
 
   test "rename table" do
@@ -1540,12 +1558,18 @@ defmodule Ecto.Adapters.TdsTest do
   end
 
   test "drop index" do
-    drop = {:drop, index(:posts, [:id], name: "posts$main")}
+    drop = {:drop, index(:posts, [:id], name: "posts$main"), :restrict}
     assert execute_ddl(drop) == [~s|DROP INDEX [posts$main] ON [posts]; |]
+
+    drop_cascade = {:drop, index(:posts, [:id], name: "posts$main"), :cascade}
+
+    assert_raise ArgumentError, ~r"MSSQL does not support `CASCADE` in DROP INDEX commands", fn ->
+      execute_ddl(drop_cascade)
+    end
   end
 
   test "drop index with prefix" do
-    drop = {:drop, index(:posts, [:id], name: "posts_category_id_permalink_index", prefix: :foo)}
+    drop = {:drop, index(:posts, [:id], name: "posts_category_id_permalink_index", prefix: :foo), :restrict}
 
     assert execute_ddl(drop) ==
              [~s|DROP INDEX [posts_category_id_permalink_index] ON [foo].[posts]; |]
@@ -1554,7 +1578,7 @@ defmodule Ecto.Adapters.TdsTest do
   test "drop index with prefix if exists" do
     drop =
       {:drop_if_exists,
-       index(:posts, [:id], name: "posts_category_id_permalink_index", prefix: :foo)}
+       index(:posts, [:id], name: "posts_category_id_permalink_index", prefix: :foo), :restrict}
 
     assert execute_ddl(drop) ==
              [
@@ -1567,10 +1591,18 @@ defmodule Ecto.Adapters.TdsTest do
                |> remove_newlines
                |> Kernel.<>(" ")
              ]
+
+    drop_cascade =
+      {:drop_if_exists,
+      index(:posts, [:id], name: "posts_category_id_permalink_index", prefix: :foo), :cascade}
+
+    assert_raise ArgumentError, ~r"MSSQL does not support `CASCADE` in DROP INDEX commands", fn ->
+      execute_ddl(drop_cascade)
+    end
   end
 
   test "drop index asserting concurrency" do
-    drop = {:drop, index(:posts, [:id], name: "posts$main", concurrently: true)}
+    drop = {:drop, index(:posts, [:id], name: "posts$main", concurrently: true), :restrict}
     assert execute_ddl(drop) == [~s|DROP INDEX [posts$main] ON [posts] LOCK=NONE; |]
   end